requirements.txt
click==6.6
ordereddict==1.1
pymongo==3.2.2
python-dotenv==0.5.1
requests==2.10.0
